+++ src/main/java/com/example/namo2/domain/moim/application/impl/MoimScheduleAndUserService.java
@@ -10,8 +10,11 @@ import com.example.namo2.global.common.exception.BaseException;
 import com.example.namo2.global.common.response.BaseResponseStatus;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Slf4j
@@ -44,6 +47,14 @@ public class MoimScheduleAndUserService {
 
     public void createMoimScheduleAlarm(MoimScheduleAlarm moimScheduleAlarm) {
         moimScheduleAlarmRepository.save(moimScheduleAlarm);
+    }
+
+    /**
+     * 메서드 네이밍에 더 분명한 의미를 담기 위해서 규칙을 좀 더
+     * 자세히 세워보는 것도 괜찮을 것 같아요
+     */
+    public List<MoimScheduleAndUser> getMoimScheduleAndUsersForMonthMoimMemo(User user, List<LocalDateTime> dates, Pageable page) {
+        return moimScheduleAndUserRepository.findMoimScheduleMemoByMonthPaging(user, dates, page);
     }
 
 }
