+++ src/main/java/com/example/namo2/domain/moim/dao/repository/MoimScheduleAndUserRepositoryCustom.java
@@ -2,10 +2,14 @@ package com.example.namo2.domain.moim.dao.repository;
 
 import com.example.namo2.domain.moim.domain.MoimScheduleAndUser;
 import com.example.namo2.domain.user.domain.User;
+import org.springframework.data.domain.Pageable;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 public interface MoimScheduleAndUserRepositoryCustom {
     List<MoimScheduleAndUser> findMoimScheduleAndUserWithMoimScheduleByUsersAndDates(LocalDateTime startDate, LocalDateTime endDate, List<User> users);
+
+    List<MoimScheduleAndUser> findMoimScheduleMemoByMonthPaging(User user, List<LocalDateTime> dates,
+                                                                Pageable pageable);
 }
