+++ src/main/java/com/example/namo2/domain/schedule/ScheduleRepositoryCustom.java
@@ -1,12 +1,12 @@
 package com.example.namo2.domain.schedule;
 
+import com.example.namo2.domain.memo.application.converter.MoimMemoResponseConverter;
+import com.example.namo2.domain.memo.ui.dto.MoimMemoResponse;
 import com.example.namo2.domain.moim.ui.dto.MoimScheduleResponse;
 import com.example.namo2.domain.schedule.domain.Schedule;
 import com.example.namo2.domain.user.domain.User;
-import com.example.namo2.domain.schedule.dto.DiaryDto;
 import com.example.namo2.domain.schedule.dto.GetScheduleRes;
 import com.example.namo2.domain.schedule.dto.OnlyDiaryDto;
-import com.example.namo2.domain.schedule.dto.SliceDiaryDto;
 import org.springframework.data.domain.Pageable;
 
 import java.time.LocalDateTime;
@@ -15,7 +15,7 @@ import java.util.List;
 public interface ScheduleRepositoryCustom {
     List<GetScheduleRes> findSchedulesByUserId(User user, LocalDateTime startDate, LocalDateTime endDate);
 
-    SliceDiaryDto<DiaryDto> findScheduleDiaryByMonthDto(User user, LocalDateTime startDate, LocalDateTime endTime, Pageable pageable);
+    MoimMemoResponse.SliceDiaryDto<MoimMemoResponse.DiaryDto> findScheduleDiaryByMonthDto(User user, LocalDateTime startDate, LocalDateTime endTime, Pageable pageable);
 
     List<OnlyDiaryDto> findAllScheduleDiary(User user);
